refactor: migrate container components to TypeScript
